// Part-9/patientor-backend/src/data/patients.ts
import { Patient, Gender, HealthCheckRating } from '../types';

const patients: Patient[] = [
  {
    id: 'd2773336-f723-11e9-8f0b-362b9e155667',
    name: 'John McClane',
    dateOfBirth: '1986-07-09',
    ssn: '090786-122X',
    gender: Gender.Male,
    occupation: 'New york city cop',
    entries: [
      {
        id: 'd811e46d-70b3-4d90-b090-4535c7cf8fb1',
        type: 'HealthCheck',
        date: '2019-08-05',
        specialist: 'MD House',
        description: 'Yearly control visit',
        healthCheckRating: HealthCheckRating.Healthy,
      },
    ],
  },
  {
    id: 'd2773598-f723-11e9-8f0b-362b9e155667',
    name: 'Martin Riggs',
    dateOfBirth: '1979-01-30',
    ssn: '300179-77A',
    gender: Gender.Male,
    occupation: 'Cop',
    entries: [
      {
        id: 'fcd59fa6-c4b4-4fec-ac4d-df4fe1f85f62',
        type: 'OccupationalHealthcare',
        date: '2019-08-05',
        specialist: 'MD House',
        employerName: 'HiQ',
        description: 'Occupational visit',
        diagnosisCodes: ['Z57.1', 'Z74.3'],
        sickLeave: { startDate: '2019-08-05', endDate: '2019-08-28' },
      },
    ],
  },
  {
    id: 'd27736ec-f723-11e9-8f0b-362b9e155667',
    name: 'Hans Gruber',
    dateOfBirth: '1970-04-25',
    ssn: '250470-555L',
    gender: Gender.Male,
    occupation: 'Technician',
    entries: [],
  },
  {
    id: 'd2773822-f723-11e9-8f0b-362b9e155667',
    name: 'Dana Scully',
    dateOfBirth: '1974-01-05',
    ssn: '050174-432N',
    gender: Gender.Female,
    occupation: 'Forensic Pathologist',
    entries: [
      {
        id: 'b4f4eca1-2aa7-4b13-9a18-4a5535c3c8da',
        type: 'Hospital',
        date: '2019-10-20',
        specialist: 'MD House',
        description: 'Hospitalization',
        diagnosisCodes: ['S62.5'],
        discharge: { date: '2019-11-25', criteria: 'Thumb has healed' },
      },
    ],
  },
  {
    id: 'd2773c6e-f723-11e9-8f0b-362b9e155667',
    name: 'Matti Luukkainen',
    dateOfBirth: '1971-04-09',
    ssn: '090471-8890',
    gender: Gender.Male,
    occupation: 'Digital evangelist',
    entries: [],
  },
];

export default patients;
